% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bin_args}
\alias{bin_args}
\title{Example of proper arguments}
\format{
A list with 4 variables, of which \code{MW.wrong}, \code{MR.wrong}, \code{MY.wrong}, \code{MZ.wrong} are all \code{FALSE}.
}
\usage{
bin_args
}
\description{
A dataset as an example containing arguments and fitting the requirements of \code{est_func()}.
}
\keyword{datasets}
