% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bin_data}
\alias{bin_data}
\title{Example of proper dataset}
\format{
A data frame with 50 rows and 14 variables:
\describe{
\item{A}{the treatment arm}
\item{Z}{the negative control exposure}
\item{W}{the negative control outcome}
\item{Y}{the outcome}
\item{V2-V10}{known covariates}
}
}
\usage{
bin_data
}
\description{
A dataset as an example containing randomly generated numbers and fitting the requirements of \code{est_func()}.
}
\keyword{datasets}
