% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_func.R
\name{est_func}
\alias{est_func}
\title{Function for estimation}
\usage{
est_func(data, arg = list(MW.wrong, MR.wrong, MY.wrong, MZ.wrong))
}
\arguments{
\item{data}{The original data. The data must contain columns \code{A}, \code{Z}, \code{W}, \code{Y}. \code{A} and \code{Y} respectively give the treatment arm and the outcome. An auxiliary exposure variable \code{Z} and an auxiliary outcome variable \code{W} should be provided as the double negative control variables.}

\item{arg}{A list of arguments.}
}
\value{
The estimation result, an object of class "proximalr_result", inheriting from the base-type "list". The members include three semiparametric estimators \code{GEST}\eqn{=\hat\Delta_1}, \code{IPW}\eqn{=\hat\Delta_2}, and \code{OR}\eqn{=\hat\Delta_3}, which operate under \eqn{\mathcal{M}_1, \mathcal{M}_2, \mathcal{M}_3}, respectively, the plug-in estimator discussed in Section 2.2.1 which we refer to as the MLE estimator hereafter, and the multiply robust (MR) estimator \code{MR}=\eqn{\hat\Delta_{\text{mr}}}.
}
\description{
The document of this function is yet to be written.
}
\examples{
est_func() # TODO

}
