% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_func.R
\name{est_func}
\alias{est_func}
\title{Function for estimation}
\usage{
est_func(
  orig.data,
  indices,
  arg = list(MW.wrong = MW.wrong, MR.wrong = MR.wrong, MY.wrong = MY.wrong, MZ.wrong =
    MZ.wrong, ATE = ATE)
)
}
\arguments{
\item{orig.data}{original data}

\item{indices}{indices}

\item{arg}{a list of arguments}
}
\value{
estimation result
}
\description{
The document of this function is yet to be written.
}
\examples{
est_func() // TODO

}
